added App.js fetch data to independent file
